src/django/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-cors-headers
djangorestframework
kavenegar
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
kavenegar==1.1.2
psycopg2>=2.8.6
sqlparse>=0.4.1
django-model-utils==4.1.1
drf_spectacular
unidecode
django-bootstrap4==3.0.1
